+++ src/querybuilder/operators.jl
@@ -1,0 +1,2 @@
+# SQL operators module
+# Add your operator functions here
